mine-safety-backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0
pillow==10.2.0
opencv-python==4.9.0.80
numpy==1.26.3
python-dotenv==1.0.0
cryptography==41.0.7

# Database drivers
psycopg2-binary==2.9.9
pymysql==1.1.0

# ML/AI
ultralytics>=8.0.0
onnx>=1.15.0
torch>=2.0.0
